==> Part2_Imputation.ipynb <==
# Implementation of all the Imputation Techniques including FPCKNN

## Doing Imputation on Multivariate Time Series Data (3D)

### Reading the Data built by the Previous Notebooks

# Loading the Raw Data
import warnings
warnings.filterwarnings("ignore")

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/1_Raw/"
raw_data = []

num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + ".pkl", 'rb') as f:
        raw_data.append(pickle.load(f))

### Nextvalue Imputation Technique

import pandas as pd
import numpy as np

def next_or_previous_value_imputation(vector):
    # Convert the vector to a Pandas Series
    series = pd.Series(vector)
    
    # Perform next value imputation (backfill)
    next_imputed_series = series[::-1].fillna(method='ffill')[::-1]

    # Perform previous value imputation (forward fill) only for NaNs at the end
    previous_imputed_series = next_imputed_series.fillna(method='ffill')

    return previous_imputed_series

# Inter Column and Between Instance Imputation

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_2_BaselineImputation/"

from tqdm import tqdm

k = 100
number_of_partitions = 5
for i in range(0,number_of_partitions):
    new_partition = np.zeros((60,25,np.array(raw_data[i]).shape[2]))
    new_partition = np.array(raw_data[i])
    
    with tqdm(np.array(raw_data[i]).shape[2]) as pbar:
        for j in range(0,np.array(raw_data[i]).shape[2]):
            new_column = np.zeros((60,25))  
            new_column = new_partition[:,:,j]
            new_column[new_column == 0.0] = np.nan
            for m in range(0,24):

                if np.isnan(new_column[:,m+1]).all():
                    new_column[:,m+1] = np.ones(60) 
                else:
                    new_column[:,m+1] = next_or_previous_value_imputation(new_column[:,m+1])
                 
                
            new_partition[:,:,j] = new_column

            pbar.update(1)

        with open(data_dir + "Partition" + str(i+1) + "_NextvalueImputation" + ".pkl", 'wb') as f:
            pickle.dump(new_partition, f)
            
            
# Between Instance Imputation

### Mean Imputation Technique

# Inter Column and Between Instance Imputation

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_2_BaselineImputation/"

from tqdm import tqdm

number_of_partitions = 5
for i in range(0,number_of_partitions):
    new_partition = np.zeros((60,25,np.array(raw_data[i]).shape[2]))
    new_partition = np.array(raw_data[i])
    
    with tqdm(np.array(raw_data[i]).shape[2]) as pbar:
        for j in range(0,np.array(raw_data[i]).shape[2]):
            new_column = np.zeros((60,25))  
            new_column = new_partition[:,:,j]
            new_column[new_column == 0.0] = np.nan
            for m in range(0,24):

                if np.isnan(new_column[:,m+1]).all():
                    new_column[:,m+1] = np.ones(60) 
                else:
                    mean_non_zero = np.mean(new_column[:,m+1][~np.isnan(new_column[:,m+1])])
                    new_column[:,m+1][np.isnan(new_column[:,m+1])] = mean_non_zero
                 
                
            new_partition[:,:,j] = new_column

            pbar.update(1)

        with open(data_dir + "Partition" + str(i+1) + "_MeanImputation" + ".pkl", 'wb') as f:
            pickle.dump(new_partition, f)
            
            
# Between Instance Imputation

# FPCKNN Imputation Technique (our Novelty)

# Correlation Function
from scipy.stats import pearsonr

def mvts_pearson_correlation(X, Y):
    if X.shape != Y.shape:
        raise ValueError("Input arrays X and Y must have the same shape.")

    # Calculate the Pearson correlation coefficient between X and Y
    X = X.T.flatten()
    Y = Y.T.flatten()
    try:
        correlation_coefficient , p = pearsonr(X, Y)
        if np.isnan(correlation_coefficient):
            correlation_coefficient = -1.0
    except:
        correlation_coefficient = -1.0
    
    return correlation_coefficient

# Inter Column and Between Instance Imputation
import warnings
warnings.filterwarnings("ignore")

from sklearn.impute import KNNImputer
imputer = KNNImputer(n_neighbors=10)

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_KnnImputation/"

from tqdm import tqdm

k_count = 50
k = 50
number_of_partitions = 5
num_attributes = 25
num_timestamps = 60

for i in range(0,number_of_partitions):
    new_partition = np.zeros((num_timestamps,num_attributes,np.array(raw_data[i]).shape[2]))
    new_partition = np.array(raw_data[i])
    
    with tqdm(new_partition.shape[2]) as pbar:
        for j in range(0,new_partition.shape[2]):
            new_column = np.zeros((num_timestamps,num_attributes))  
            new_column = new_partition[:,:,j]
            
            
            
            new_column[new_column == 0.] = np.nan
            
            if np.isnan(new_column[:,1:25]).all():
                new_column = new_partition[:,:,j-1]
            
            nan_index = []
            if np.isnan(new_column).any():
                for m in range(0,num_attributes-1):
                    if np.isnan(new_column[:,m+1]).any():
                        nan_index.append(m+1)
                
                if j < k_count:
                    if j == 0:
                        k = 1
                    else:
                        k = j
                else:
                    k = k_count
                correlation_coefficient = np.full(k, -2.0)
                the_X = new_column
                
                the_X = np.nan_to_num(the_X, nan=0.0)
                
                for n in range(0, k):
                    the_Y = new_partition[:,:,j-n-1]
                    correlation_coefficient[n] = mvts_pearson_correlation(the_X[:,1:25], the_Y[:,1:25])
                    
                for m in range(0,num_attributes-1):

                    if np.isnan(new_column[:,m+1]).all():
                        
                        indices_of_largest = np.where(correlation_coefficient == np.max(correlation_coefficient))
                        first_occurrence_index = indices_of_largest[0][0]
                        new_column[:,m+1] = new_partition[:,m+1,j-first_occurrence_index-1]

                    else:
                        if j>=2:

                            sorted_indices = np.argsort(correlation_coefficient)[::-1]  # Sort in descending order
                            largest_index = sorted_indices[0]  # Index of the largest item
                            second_largest_index = sorted_indices[1]

                            new_2d = [new_partition[:,m+1,j-second_largest_index-1], new_partition[:,m+1,j-largest_index-1], new_column[:,m+1]]
                            new_column[:,m+1] = imputer.fit_transform(new_2d)[2,:]
                        else:
                            new_2d = new_column[:,m+1].reshape(-1, 1)
                            new_column[:,m+1] = imputer.fit_transform(new_2d)[:,0]
                
                
            new_partition[:,:,j] = new_column

            pbar.update(1)

        with open(data_dir + "Partition" + str(i+1) + "_KnnImputation" + ".pkl", 'wb') as f:
            pickle.dump(new_partition, f)
            
            
# Between Instance Imputation

### Missing Value Exploration

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_KnnImputation/"
imputed_data = []

number_of_partitions = 5
for i in range(1,number_of_partitions +1):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i) + "_KnnImputation" + ".pkl", 'rb') as f:
        imputed_data.append(pickle.load(f))

def print_missing_values(data, start_partition, end_partition):
    abt_header = ['Timestamp', 'R_VALUE','TOTUSJH','TOTBSQ','TOTPOT','TOTUSJZ','ABSNJZH','SAVNCPP',
                               'USFLUX','TOTFZ','MEANPOT', 'EPSX', 'EPSY','EPSZ','MEANSHR','SHRGT45','MEANGAM',
                                  'MEANGBT','MEANGBZ','MEANGBH','MEANJZH','TOTFY','MEANJZD','MEANALP','TOTFX']
    num_columns = 25
    num_timestamps = 60
    num_partitions = 5
    null_count = [0,0,0,0,0]
    non_null_count = [0,0,0,0,0]
    null_count_per_feature = np.zeros((num_partitions,num_columns), dtype=int)

    for i in range(start_partition-1, end_partition):
        partition = np.array(data[i])

        for j in range(0,partition.shape[2]):
            mvts = partition[:,:, j]
            for m in range(0,num_columns):
                for n in range (0,num_timestamps):
                    if (mvts[n,m] == 0.0 or np.isnan(mvts[n,m]) or np.isinf(mvts[n,m])):
                        null_count[i] += 1
                        null_count_per_feature[i,m] += 1
                    else:
                        non_null_count[i] += 1

        print("Partition" + str(i+1) + ":")
        print("null counts in P" + str(i+1) + ": " + str(null_count[i]))
        print("non-null counts in P"+ str(i+1) + ": " + str(non_null_count[i]))
        for x in range(0,num_columns):
            print(abt_header[x] + ": " + str(null_count_per_feature[i,x]))

        print("\n")

def print_values(data, start_partition, end_partition):
    abt_header = ['Timestamp', 'R_VALUE','TOTUSJH','TOTBSQ','TOTPOT','TOTUSJZ','ABSNJZH','SAVNCPP',
                               'USFLUX','TOTFZ','MEANPOT', 'EPSX', 'EPSY','EPSZ','MEANSHR','SHRGT45','MEANGAM',
                                  'MEANGBT','MEANGBZ','MEANGBH','MEANJZH','TOTFY','MEANJZD','MEANALP','TOTFX']
    num_columns = 2
    num_partitions = 5

    for i in range(start_partition-1, end_partition):
        partition = np.array(data[i])
        print('Partition: ' + str(i+1))
        for j in range(1,partition.shape[1]):
            mvts = partition[:,j, :]
            print('Min ' + abt_header[j] + ': ' + str(np.min(mvts)))
            print('Max ' + abt_header[j] + ': ' + str(np.max(mvts)))

            print("\n")

print_values(imputed_data,1,5)

### Saving Final Data Imputed by FPCKNN

We chose to name our imputation technique KnnImputation instead of FPCKNN for saving as a file, as it more clearly indicates that it is based on the KNN technique.

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_KnnImputation/"
imputed_data = []

number_of_partitions = 5
for i in range(1,number_of_partitions +1):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i) + "_KnnImputation" + ".pkl", 'rb') as f:
        imputed_data.append(pickle.load(f))

import pandas as pd
labels = []
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/2_Labels/"

for i in range(1,6):
    labels.append(pd.read_csv(data_dir + "Partition" + str(i) + "_labels.csv"))

## Doing Imputation on Concatenated Multivariate Time Series Data (2D)

### Converting the 3D Multivariate Time Series Data to a 2D Data

# Concatenation with Shuffle

def multi_to_uni(start_partition, end_partition, data_dir, data, labels, name):
    category_mapping = {'X': 1, 'M': 1, 'B': 0, 'C': 0, 'FQ': 0}
    
    num_attributes = 25
    num_timestamps = 60
    
    for i in range(start_partition-1,end_partition):
        
        the_labels = pd.DataFrame()
        the_labels['FLARE_CLASS'] = labels[i]['FLARE_CLASS'].map(category_mapping)
        new_partition = np.zeros((np.array(data[i]).shape[2], num_timestamps*(num_attributes-1)))
        new_partition_label = np.zeros(new_partition.shape[0])
        
        each_partition = np.zeros((num_timestamps, num_attributes, new_partition.shape[0]))
        each_partition = np.array(data[i])
        
        with tqdm(new_partition.shape[0]) as pbar:
            for j in range(0,new_partition.shape[0]):
                new_column = np.zeros((num_timestamps,num_attributes)) 
                new_column = each_partition[:,:,j]

                flettened = np.zeros(num_timestamps*(num_attributes-1))

                for m in range(1,num_attributes):
                    flettened[(m-1)*num_timestamps:m*num_timestamps] = new_column[:,m]

                new_partition[j,:] = flettened
                new_partition_label[j] = the_labels.iloc[j]
                
                
                pbar.update(1)
        
        print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(new_partition).any()))
        X_train = new_partition
        Y_train = new_partition_label


        num_samples = X_train.shape[0]
        shuffle_indices = np.random.permutation(num_samples)

        X_train_shuffled = X_train[shuffle_indices]
        Y_train_shuffled = Y_train[shuffle_indices]
    

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Concatenation_" + name + ".pkl", 'wb') as f:
            pickle.dump(X_train_shuffled, f)

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Labels_Concatenation_" + name + ".pkl", 'wb') as f:
            pickle.dump(Y_train_shuffled, f)

from tqdm import tqdm
import pandas as pd
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_1_FinalData_Concatenation_KnnImputation/"

multi_to_uni(1, 5, data_dir, imputed_data, labels)

### Mean Imputation Technique

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_2_BaselineImputation/"
imputed_data = []

number_of_partitions = 5
for i in range(1,number_of_partitions +1):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i) + "_MeanImputation" + ".pkl", 'rb') as f:
        imputed_data.append(pickle.load(f))

import pandas as pd
labels = []
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/2_Labels/"

for i in range(1,6):
    labels.append(pd.read_csv(data_dir + "Partition" + str(i) + "_labels.csv"))

from tqdm import tqdm
import pandas as pd
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"
name = 'MeanImputation'
multi_to_uni(1, 5, data_dir, imputed_data, labels, name)

### Nextvalue Imputation Technique

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/3_2_BaselineImputation/"
imputed_data = []

number_of_partitions = 5
for i in range(1,number_of_partitions +1):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i) + "_NextvalueImputation" + ".pkl", 'rb') as f:
        imputed_data.append(pickle.load(f))

import pandas as pd
labels = []
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/2_Labels/"

for i in range(1,6):
    labels.append(pd.read_csv(data_dir + "Partition" + str(i) + "_labels.csv"))

from tqdm import tqdm
import pandas as pd
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"
name = 'NextvalueImputation'
multi_to_uni(1, 5, data_dir, imputed_data, labels, name)

## Doing Imputation on Vector Data

For each time series feature of SWAN-SF, we calculate nine statistical features to convert the multivariate time series data into vector data.

# NewFeatures with Shuffle

def new_Features_pkl(start_partition, end_partition, data_dir, data, labels, name):
    category_mapping = {'X': 1, 'M': 1, 'B': 0, 'C': 0, 'FQ': 0}

    number_of_new_features = 9
    num_attributes = 25
    num_timestamps = 60
    
    for i in range(start_partition-1,end_partition):
        
        the_labels = pd.DataFrame()
        the_labels['FLARE_CLASS'] = labels[i]['FLARE_CLASS'].map(category_mapping)
        new_partition = np.zeros((np.array(data[i]).shape[2], number_of_new_features*24))
        new_partition_label = np.zeros(new_partition.shape[0])
        
        each_partition = np.zeros((num_timestamps, num_attributes, new_partition.shape[0]))
        each_partition = np.array(data[i])
        
        with tqdm(new_partition.shape[0]) as pbar:
            for j in range(0,new_partition.shape[0]):
                new_column = np.zeros((num_timestamps,num_attributes)) 
                new_column = each_partition[:,:,j]
                

                new_features = np.zeros(number_of_new_features*(num_attributes-1))

                for m in range(1,num_attributes):
                    
                    mean = np.mean(new_column[:,m])
                    new_features[((m-1)*number_of_new_features) + 0] = mean
                    median = np.median(new_column[:,m])
                    new_features[((m-1)*number_of_new_features) + 1] = median
                    std = np.std(new_column[:,m])
                    new_features[((m-1)*number_of_new_features) + 2] = std
                    
                    skewness = skew(new_column[:,m])
                    if np.isnan(skewness):
                        skewness = new_partition[j-1, ((m-1)*number_of_new_features) + 3]
                    new_features[((m-1)*number_of_new_features) + 3] = skewness
                    
                    kurtosis_value = kurtosis(new_column[:,m])
                    if np.isnan(kurtosis_value):
                        kurtosis_value = new_partition[j-1, ((m-1)*number_of_new_features) + 4]
                    new_features[((m-1)*number_of_new_features) + 4] = kurtosis_value
                    
                    indices = np.arange(num_timestamps)
                    weight_array = indices / num_timestamps
                    weighted_avg = np.average(new_column[:,m], weights=weight_array)
                    if np.isnan(weighted_avg):
                        weighted_avg = new_partition[j-1, ((m-1)*number_of_new_features) + 5]
                    new_features[((m-1)*number_of_new_features) + 5] = weighted_avg
                    
                    last_value = new_column[59,m]
                    new_features[((m-1)*number_of_new_features) + 6] = last_value
                    first_value = new_column[0,m]
                    new_features[((m-1)*number_of_new_features) + 7] = first_value
                    
                    numerator = np.sum((new_column[:,m] - mean) * (indices - np.mean(indices)))
                    denominator = np.sum((new_column[:,m] - mean) ** 2)
                    slope = numerator / denominator
                    if np.isnan(slope):
                        slope = new_partition[j-1, ((m-1)*number_of_new_features) + 8]
                    new_features[((m-1)*number_of_new_features) + 8] = slope
                    
                    if all(value == 1.0 for value in new_column[:,m]):
                        new_features[((m-1)*number_of_new_features) + 0 : ((m-1)*number_of_new_features) + 9] = np.ones(9)
                
                
                if np.isnan(new_features).any():
                    print('nan')
                    print()
                    mean_non_zero = np.mean(new_features[~np.isnan(new_features)])
                    new_features[np.isnan(new_features)] = mean_non_zero    
                    
                new_partition[j,:] = new_features
                new_partition_label[j] = the_labels.iloc[j]
                
                pbar.update(1)
                

        print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(new_partition).any()))  
        X_train = new_partition
        Y_train = new_partition_label


        num_samples = X_train.shape[0]
        shuffle_indices = np.random.permutation(num_samples)

        X_train_shuffled = X_train[shuffle_indices]
        Y_train_shuffled = Y_train[shuffle_indices]
    

        with open(data_dir + "Partition" + str(i+1) 
                       + "_NewFeatures_" + name + ".pkl", 'wb') as f:
            pickle.dump(X_train_shuffled, f)

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Labels_NewFeatures_" + name + ".pkl", 'wb') as f:
            pickle.dump(Y_train_shuffled, f)

from tqdm import tqdm
import pandas as pd
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_2_FinalData_NewFeatures_KnnImputation/"

new_Features_pkl(1, 5, data_dir, imputed_data, labels)

### Mean Imputation Technique

from tqdm import tqdm
import pandas as pd
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"
name = 'MeanImputation'
new_Features_pkl(1, 5, data_dir, imputed_data, labels, name)

### Nextvalue Imputation Technique

from tqdm import tqdm
import pandas as pd
from scipy.stats import skew, kurtosis
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"
name = 'NextvalueImputation'
new_Features_pkl(1, 5, data_dir, imputed_data, labels, name)

## Final Data Without Imputation

### Concatenated Multivariate Time Series Data (2D)

# Concatenation with Shuffle

def multi_to_uni(start_partition, end_partition, data_dir, data, labels):
    category_mapping = {'X': 1, 'M': 1, 'B': 0, 'C': 0, 'FQ': 0}
    
    num_attributes = 25
    num_timestamps = 60
    
    for i in range(start_partition-1,end_partition):
        
        the_labels = pd.DataFrame()
        the_labels['FLARE_CLASS'] = labels[i]['FLARE_CLASS'].map(category_mapping)
        new_partition = np.zeros((np.array(data[i]).shape[2], num_timestamps*(num_attributes-1)))
        new_partition_label = np.zeros(new_partition.shape[0])
        
        each_partition = np.zeros((num_timestamps, num_attributes, new_partition.shape[0]))
        each_partition = np.array(data[i])
        
        with tqdm(new_partition.shape[0]) as pbar:
            for j in range(0,new_partition.shape[0]):
                new_column = np.zeros((num_timestamps,num_attributes)) 
                new_column = each_partition[:,:,j]
                
                new_column = np.nan_to_num(new_column, nan=0.0)
                
                flettened = np.zeros(num_timestamps*(num_attributes-1))

                for m in range(1,num_attributes):
                    flettened[(m-1)*num_timestamps:m*num_timestamps] = new_column[:,m]

                new_partition[j,:] = flettened
                new_partition_label[j] = the_labels.iloc[j]
                
                
                pbar.update(1)
        
        print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(new_partition).any()))
        X_train = new_partition
        Y_train = new_partition_label


        num_samples = X_train.shape[0]
        shuffle_indices = np.random.permutation(num_samples)

        X_train_shuffled = X_train[shuffle_indices]
        Y_train_shuffled = Y_train[shuffle_indices]
    

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Concatenation_Raw" + ".pkl", 'wb') as f:
            pickle.dump(X_train_shuffled, f)

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Labels_Concatenation_Raw" + ".pkl", 'wb') as f:
            pickle.dump(Y_train_shuffled, f)

from tqdm import tqdm
import pandas as pd
data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_6_FinalData_Concatenation_Raw/"

multi_to_uni(1, 5, data_dir, raw_data, labels)

### Vector Data

# NewFeatures with Shuffle

def new_Features_pkl(start_partition, end_partition, data_dir, data, labels):
    category_mapping = {'X': 1, 'M': 1, 'B': 0, 'C': 0, 'FQ': 0}

    number_of_new_features = 9
    num_attributes = 25
    num_timestamps = 60
    
    for i in range(start_partition-1,end_partition):
        
        the_labels = pd.DataFrame()
        the_labels['FLARE_CLASS'] = labels[i]['FLARE_CLASS'].map(category_mapping)
        new_partition = np.zeros((np.array(data[i]).shape[2], number_of_new_features*24))
        new_partition_label = np.zeros(new_partition.shape[0])
        
        each_partition = np.zeros((num_timestamps, num_attributes, new_partition.shape[0]))
        each_partition = np.array(data[i])
        
        with tqdm(new_partition.shape[0]) as pbar:
            for j in range(0,new_partition.shape[0]):
                new_column = np.zeros((num_timestamps,num_attributes)) 
                new_column = each_partition[:,:,j]
                
                new_column = np.nan_to_num(new_column, nan=0.0)

                new_features = np.zeros(number_of_new_features*(num_attributes-1))

                for m in range(1,num_attributes):
                    all_zeros = np.all(new_column[:,m] == 0.0)
                    if all_zeros:
                        new_features[((m-1)*number_of_new_features):((m-1)*number_of_new_features)+9] = 0
                    else:
                        mean = np.mean(new_column[:,m])
                        new_features[((m-1)*number_of_new_features) + 0] = mean
                        median = np.median(new_column[:,m])
                        new_features[((m-1)*number_of_new_features) + 1] = median
                        std = np.std(new_column[:,m])
                        new_features[((m-1)*number_of_new_features) + 2] = std

                        skewness = skew(new_column[:,m])
                        if np.isreal(skewness) == False:
                            skewness = new_partition[j-1, ((m-1)*number_of_new_features) + 3]
                        new_features[((m-1)*number_of_new_features) + 3] = skewness

                        kurtosis_value = kurtosis(new_column[:,m])
                        if np.isreal(kurtosis_value) == False:
                            kurtosis_value = new_partition[j-1, ((m-1)*number_of_new_features) + 4]
                        new_features[((m-1)*number_of_new_features) + 4] = kurtosis_value

                        indices = np.arange(num_timestamps)
                        weight_array = indices / num_timestamps
                        weighted_avg = np.average(new_column[:,m], weights=weight_array)
                        if weighted_avg == np.nan:
                            weighted_avg = new_partition[j-1, ((m-1)*number_of_new_features) + 5]
                        new_features[((m-1)*number_of_new_features) + 5] = weighted_avg

                        last_value = new_column[59,m]
                        new_features[((m-1)*number_of_new_features) + 6] = last_value
                        first_value = new_column[0,m]
                        new_features[((m-1)*number_of_new_features) + 7] = first_value

                        numerator = np.sum((new_column[:,m] - mean) * (indices - np.mean(indices)))
                        denominator = np.sum((new_column[:,m] - mean) ** 2)
                        slope = numerator / denominator
                        if np.isreal(slope) == False:
                            slope = new_partition[j-1, ((m-1)*number_of_new_features) + 8]
                        new_features[((m-1)*number_of_new_features) + 8] = slope
                
                new_partition[j,:] = new_features
                new_partition_label[j] = the_labels.iloc[j]
                
                pbar.update(1)
                

        print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(new_partition).any()))  
        X_train = new_partition
        Y_train = new_partition_label


        num_samples = X_train.shape[0]
        shuffle_indices = np.random.permutation(num_samples)

        X_train_shuffled = X_train[shuffle_indices]
        Y_train_shuffled = Y_train[shuffle_indices]
    

        with open(data_dir + "Partition" + str(i+1) 
                       + "_NewFeatures_Raw" + ".pkl", 'wb') as f:
            pickle.dump(X_train_shuffled, f)

        with open(data_dir + "Partition" + str(i+1) 
                       + "_Labels_NewFeatures_Raw" + ".pkl", 'wb') as f:
            pickle.dump(Y_train_shuffled, f)

from tqdm import tqdm
import pandas as pd
from scipy.stats import skew, kurtosis

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_7_FinalData_NewFeatures_Raw/"

new_Features_pkl(1, 5, data_dir, raw_data, labels)
