==> Part4_Normalization.ipynb <==
# Normalization on Vector Data

### Reading the Data built by the Previous Notebooks (FPCKNN Imputation)

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_2_FinalData_NewFeatures_KnnImputation/"

newf_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_KnnImputation" +".pkl", 'rb') as f:
        newf_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(newf_imputed_data[i]).any() or np.isinf(newf_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_KnnImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (Mean Imputation)

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"

newf_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_MeanImputation" +".pkl", 'rb') as f:
        newf_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(newf_imputed_data[i]).any() or np.isinf(newf_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_MeanImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (Nextvalue Imputation)

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"

newf_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_NextvalueImputation" +".pkl", 'rb') as f:
        newf_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(newf_imputed_data[i]).any() or np.isinf(newf_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_NextvalueImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Zscore Normalization Technique

from scipy import stats
import numpy as np
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from tqdm import tqdm

def zscore(start_partition, end_partition, data, labels, data_dir, name):
    
    num_new_features = 9
    num_attributes = 24
    num_all_features = num_attributes * num_new_features

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_all_features))
        new_partition = np.array(data[i])

        with tqdm(num_all_features) as pbar:
            for j in range(0,num_all_features):
                new_column = np.zeros(new_partition.shape[0])  
                new_column = new_partition[:,j]

                
                if np.std(new_column)== 0.0:
                    zscore = np.ones(new_partition.shape[0])
                else:
                    zscore = stats.zscore(new_column)
                if (np.isnan(zscore).any() or np.isinf(zscore).any()):
                    print('nan-zscore')

                new_column = zscore


                new_partition[:,j] = new_column
                pbar.update(1)

            with open(data_dir + "Partition" + str(i+1) + "_" + name +".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

        
                   

## MinMax Normalization Technique

from scipy import stats
import numpy as np
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from tqdm import tqdm

def minmax(start_partition, end_partition, data, labels, data_dir, name):
    
    num_new_features = 9
    num_attributes = 24
    num_all_features = num_attributes * num_new_features

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_all_features))
        new_partition = np.array(data[i])

        with tqdm(num_all_features) as pbar:
            for j in range(0,num_all_features):
                new_column = np.zeros(new_partition.shape[0])  
                new_column = new_partition[:,j]

                
                if np.std(new_column)== 0.0:
                    minmax = np.ones(new_partition.shape[0])
                else:
                    data_2d = new_column.reshape(-1, 1)
                    minmax = scaler.fit_transform(data_2d).flatten()
                if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                    print('nan-zscore')

                new_column = minmax


                new_partition[:,j] = new_column
                pbar.update(1)

            with open(data_dir + "Partition" + str(i+1) + "_" + name +".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

        
                   

### Saving Z-score Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_1_FinalData_NewFeatures_ZM_KnnImputation/"
name = 'NewFeatures_ZNorm_KnnImputation'

zscore(1,5,newf_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_1_FinalData_NewFeatures_ZM_KnnImputation/"
name = 'NewFeatures_MinMaxNorm_KnnImputation'

minmax(1,5,newf_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and Mean Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_MinMaxNorm_MeanImputation'

minmax(1,5,newf_imputed_data,labels, data_dir, name)

### Saving Z-score Normalization and Mean Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_ZNorm_MeanImputation'

zscore(1,5,newf_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and Nextvalue Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_MinMaxNorm_NextvalueImputation'

minmax(1,5,newf_imputed_data,labels, data_dir, name)

### Saving Z-score Normalization and Nextvalue Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_ZNorm_NextvalueImputation'

zscore(1,5,newf_imputed_data,labels, data_dir, name)

# LSBZM Normalization Technique for Vector Data (our Novelty)

import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from scipy import stats
from tqdm import tqdm

def lsbzm(start_partition, end_partition, data, labels, data_dir, name):

    num_new_features = 9
    num_attributes = 24
    num_all_features = num_attributes * num_new_features
    

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_all_features))
        new_partition = np.array(data[i])

        with tqdm(num_all_features) as pbar:
            for j in range(0,num_all_features):
                new_column = np.zeros(new_partition.shape[0])  
                new_column = new_partition[:,j]

                minmax = np.zeros(new_partition.shape[0])
                all_positive = np.zeros(new_partition.shape[0])

                the_min = np.min(new_column)
                the_max = np.max(new_column)
                skewness = stats.skew(new_column)
                
                        
                if (the_max - the_min > 100000):
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_column + 2*abs(the_min)
                        else:
                            all_positive = new_column + 0.1
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log')
                            

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_column + abs(the_min)
                        else:
                            all_positive = new_column + 0.1

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt')

                    else:
                        if np.std(new_column)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            zscore = stats.zscore(new_column)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')
                
                        
                elif (the_max < 1 and the_min > -1):
                    
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_column + 2*abs(the_min)
                        else:
                            all_positive = new_column + 0.1
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log') 

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_column + abs(the_min)
                        else:
                            all_positive = new_column + 0.1

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt') 

                    else:
                        if np.std(new_column)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            zscore = stats.zscore(new_column)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')
                            
                        
                else:
                    if (skewness > 1 or skewness < -1):
                        if (the_min < 0):
                            all_positive = new_column + 2*abs(the_min)
                        else:
                            all_positive = new_column + 0.1

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            boxcox, values = stats.boxcox(all_positive)
                            data_2d = boxcox.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-boxcox')

                            
                    else:
                        if np.std(new_column)== 0.0:
                            minmax = np.ones(new_partition.shape[0])
                        else:
                            zscore = stats.zscore(new_column)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')

                new_column = minmax

                new_partition[:,j] = new_column

                pbar.update(1)

            with open(data_dir + "Partition" + str(i+1) + "_" + name + ".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

### Saving LSBZM Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_2_FinalData_NewFeatures_LSBZM_KnnImputation/"
name = 'NewFeatures_LSBZM_KnnImputation'

lsbzm(1,5,newf_imputed_data,labels, data_dir, name)

### Saving LSBZM Normalization and Mean Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_LSBZM_MeanImputation'

lsbzm(1,5,newf_imputed_data,labels, data_dir, name)

### Saving LSBZM Normalization and Nextvalue Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'NewFeatures_LSBZM_NextvalueImputation'

lsbzm(1,5,newf_imputed_data,labels, data_dir, name)

## Data Without B and C Classes

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_6_FinalData_WithoutB&C_NewFeatures_KnnImputation/"

imputed_wBandC_data = []
wBandC_labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_WithoutB&C_KnnImputation" +".pkl", 'rb') as f:
        imputed_wBandC_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(imputed_wBandC_data[i]).any() or np.isinf(imputed_wBandC_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_WithoutB&C_KnnImputation" +".pkl", 'rb') as f:
        wBandC_labels.append(pickle.load(f))

### Saving FPCKNN Imputation Data (Without B and C Classes)

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_5_FinalData_NewFeatures_WithoutB&C_LSBZM_KnnImputation/"
name = 'NewFeatures_WithoutB&C_LSBZM_KnnImputation'

lsbzm(1,5,imputed_wBandC_data,wBandC_labels, data_dir, name)

## Data Without C Class

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_7_FinalData_WithoutC_NewFeatures_KnnImputation/"

imputed_wC_data = []
wC_labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_WithoutC_KnnImputation" +".pkl", 'rb') as f:
        imputed_wC_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(imputed_wC_data[i]).any() or np.isinf(imputed_wC_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_WithoutC_KnnImputation" +".pkl", 'rb') as f:
        wC_labels.append(pickle.load(f))

### Saving FPCKNN Imputation Data (Without C Class)

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_6_FinalData_NewFeatures_WithoutC_LSBZM_KnnImputation/"
name = 'NewFeatures_WithoutC_LSBZM_KnnImputation'

lsbzm(1,5,imputed_wC_data, wC_labels, data_dir, name)

### Missing value Exploration

# Loading the Raw Data
import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_2_FinalData_NewFeatures_LSBZM_KnnImputation/"
normalized_data = []

num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        normalized_data.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(normalized_data[i]).any() or np.isinf(normalized_data[i]).any()))  

# Loading the Raw Data
import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_1_FinalData_NewFeatures_ZM_KnnImputation/"
normalized_data = []

num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_ZNorm_KnnImputation" + ".pkl", 'rb') as f:
        normalized_data.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(normalized_data[i]).any() or np.isinf(normalized_data[i]).any()))
