==> Part5_Normalization.ipynb <==
# Normalization on Concatenated Multivariate Time Series Data (2D)

### Reading the Data built by the Previous Notebooks (FPCKNN Imputation)

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_1_FinalData_Concatenation_KnnImputation/"

concat_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        concat_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(concat_imputed_data[i]).any() or np.isinf(concat_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Mean Imputation Technique

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"

concat_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_MeanImputation" +".pkl", 'rb') as f:
        concat_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(concat_imputed_data[i]).any() or np.isinf(concat_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_Concatenation_MeanImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Nextvalue Imputation Technique

import numpy as np
import pandas as pd
import pickle

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/4_0_FinalData_BaseLineImputation/"

concat_imputed_data = []
labels = []


num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_NextvalueImputation" +".pkl", 'rb') as f:
        concat_imputed_data.append(pickle.load(f))
        
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(concat_imputed_data[i]).any() or np.isinf(concat_imputed_data[i]).any()))  

    with open(data_dir + "Partition" + str(i+1) + "_Labels_Concatenation_NextvalueImputation" +".pkl", 'rb') as f:
        labels.append(pickle.load(f))

### Zscore Normalization Technique

from scipy import stats
import numpy as np
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from tqdm import tqdm

def zscore(start_partition, end_partition, data, labels, data_dir, name):
    
    num_timestamps= 60
    num_attributes = 24

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_timestamps*num_attributes))
        new_partition = np.array(data[i])

        with tqdm(num_attributes) as pbar:
            for j in range(0,num_attributes):
                new_attribute = np.zeros(new_partition.shape[0]*num_timestamps)  
                
                for m in range(0,new_partition.shape[0]):
                    new_attribute[m*num_timestamps:(m+1)*num_timestamps] = new_partition[m,j*num_timestamps:(j+1)*num_timestamps]
                 
                if np.std(new_attribute)== 0.0:
                    zscore = np.ones(new_partition.shape[0]*num_timestamps)
                else:
                    zscore = stats.zscore(new_attribute)
                if (np.isnan(zscore).any() or np.isinf(zscore).any()):
                    print('nan-zscore')
                
                for m in range(0,new_partition.shape[0]):
                    new_partition[m,j*num_timestamps:(j+1)*num_timestamps] = zscore[m*num_timestamps:(m+1)*num_timestamps] 
                 
                pbar.update(1)

            with open(data_dir + "Partition" + str(i+1) + "_" + name +".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

        
                   

### MinMax Normalization Technique

from scipy import stats
import numpy as np
from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from tqdm import tqdm

def minmax(start_partition, end_partition, data, labels, data_dir, name):
    
    num_timestamps= 60
    num_attributes = 24

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_timestamps*num_attributes))
        new_partition = np.array(data[i])

        with tqdm(num_attributes) as pbar:
            for j in range(0,num_attributes):
                new_attribute = np.zeros(new_partition.shape[0]*num_timestamps)  
                
                for m in range(0,new_partition.shape[0]):
                    new_attribute[m*num_timestamps:(m+1)*num_timestamps] = new_partition[m,j*num_timestamps:(j+1)*num_timestamps]
                 
                if np.std(new_attribute)== 0.0:
                    minmax = np.ones(new_partition.shape[0]*num_timestamps)
                else:
                    data_2d = new_attribute.reshape(-1, 1)
                    minmax = scaler.fit_transform(data_2d).flatten()
                if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                    print('nan-minmax')
                
                for m in range(0,new_partition.shape[0]):
                    new_partition[m,j*num_timestamps:(j+1)*num_timestamps] = minmax[m*num_timestamps:(m+1)*num_timestamps] 
                 
                pbar.update(1)

            with open(data_dir + "Partition" + str(i+1) + "_" + name +".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

        
                   

### Saving Z-score Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_3_FinalData_Concatenation_ZM_KnnImputation/"
name = 'Concatenation_ZNorm_KnnImputation'

zscore(1,5,concat_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_3_FinalData_Concatenation_ZM_KnnImputation/"
name = 'Concatenation_MinMaxNorm_KnnImputation'

minmax(1,5,concat_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and Mean Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'Concatenation_MinMaxNorm_MeanImputation'

minmax(1,5,concat_imputed_data,labels, data_dir, name)

### Saving Z-score Normalization and Mean Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'Concatenation_ZNorm_MeanImputation'

zscore(1,5,concat_imputed_data,labels, data_dir, name)

### Saving MinMax Normalization and Nextvalue Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'Concatenation_MinMaxNorm_NextvalueImputation'

minmax(1,5,concat_imputed_data,labels, data_dir, name)

### Saving Z-score Normalization and Nextvalue Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
name = 'Concatenation_ZNorm_NextvalueImputation'

zscore(1,5,concat_imputed_data,labels, data_dir, name)

# LSBZM Normalization Technique for Concatenated Multivariate Time Series Data (2D) (our Novelty)

import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from scipy import stats
from tqdm import tqdm

def lsbzm(start_partition, end_partition, data, labels, data_dir, name):

    num_timestamps= 60
    num_attributes = 24

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_timestamps*num_attributes))
        new_partition = np.array(data[i])

        with tqdm(num_attributes) as pbar:
            for j in range(0,num_attributes):
                new_attribute = np.zeros(new_partition.shape[0]*num_timestamps)  
                
                for m in range(0,new_partition.shape[0]):
                    new_attribute[m*num_timestamps:(m+1)*num_timestamps] = new_partition[m,j*num_timestamps:(j+1)*num_timestamps]
                
                minmax = np.zeros(new_partition.shape[0]*num_timestamps)
                all_positive = np.zeros(new_partition.shape[0]*num_timestamps)

                the_min = np.min(new_attribute)
                the_max = np.max(new_attribute)
                skewness = stats.skew(new_attribute)
                
                if (the_max - the_min > 100000):
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log')

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt')

                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')
                
                
                        
                elif (the_max < 1 and the_min > -1):
                    
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log')

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt')

                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')

                        
                else:
                    if (skewness > 1 or skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            boxcox, values = stats.boxcox(all_positive)
                            data_2d = boxcox.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-boxcox')
                            
                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')

                    new_attribute = minmax

                for m in range(0,new_partition.shape[0]):
                    new_partition[m,j*num_timestamps:(j+1)*num_timestamps] = minmax[m*num_timestamps:(m+1)*num_timestamps] 
                 

                pbar.update(1)

       

            with open(data_dir + "Partition" + str(i+1) + "_" + name + ".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + "Partition" + str(i+1) + "_Labels_" + name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

### Saving LSBZM Normalization and FPCKNN Imputation Data

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_4_FinalData_Concatenation_LSBZM_KnnImputation/"
name = 'Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,concat_imputed_data,labels, data_dir, name)

### Missing value Exploration

# Loading the Raw Data
import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_4_FinalData_Concatenation_LSBZM_KnnImputation/"
normalized_data = []

num_partitions = 5

for i in range(0,num_partitions):
# Load the array with Pickle
    with open(data_dir + "Partition" + str(i+1) + "_Concatenation_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        normalized_data.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(normalized_data[i]).any() or np.isinf(normalized_data[i]).any()))  

### Extracting Information for Table 3 of the Paper

import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from scipy import stats
from tqdm import tqdm

def lsbzm_info(start_partition, end_partition, data, labels, data_dir, name):

    num_timestamps= 60
    num_attributes = 24

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_timestamps*num_attributes))
        new_partition = np.array(data[i])
        
        abt_header = ['R_VALUE','TOTUSJH','TOTBSQ','TOTPOT','TOTUSJZ','ABSNJZH','SAVNCPP',
                           'USFLUX','TOTFZ','MEANPOT', 'EPSX', 'EPSY','EPSZ','MEANSHR','SHRGT45','MEANGAM',
                              'MEANGBT','MEANGBZ','MEANGBH','MEANJZH','TOTFY','MEANJZD','MEANALP','TOTFX']

        for j in range(0,num_attributes):
            new_attribute = np.zeros(new_partition.shape[0]*num_timestamps)  

            for m in range(0,new_partition.shape[0]):
                new_attribute[m*num_timestamps:(m+1)*num_timestamps] = new_partition[m,j*num_timestamps:(j+1)*num_timestamps]

            minmax = np.zeros(new_partition.shape[0]*num_timestamps)
            all_positive = np.zeros(new_partition.shape[0]*num_timestamps)

            the_min = np.min(new_attribute)
            the_max = np.max(new_attribute)
            skewness = stats.skew(new_attribute)

            print(abt_header[j]+ ':')
            
            if (the_max - the_min > 100000):
                print('large-values')
                if (skewness > 1):
                    print('right-skewness')
                    if (the_min < 0):
                        print('negative-values')
                        all_positive = new_attribute + 2*abs(the_min)
                    else:
                        all_positive = new_attribute

                    if np.std(all_positive)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:log')
                        log = np.log(all_positive)
                        data_2d = log.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-log')

                elif (skewness < -1):
                    print('left-skewness')
                    if (the_min < 0):
                        print('negative-values')
                        all_positive = new_attribute + abs(the_min)
                    else:
                        all_positive = new_attribute

                    if np.std(all_positive)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:sqrt')
                        sqrt = np.sqrt(all_positive)
                        data_2d = sqrt.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-sqrt')

                else:
                    if np.std(new_attribute)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:zscore')
                        zscore = stats.zscore(new_attribute)
                        data_2d = zscore.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-zscore')



            elif (the_max < 1 and the_min > -1):
                print('small-value')
                if (skewness > 1):
                    print('right-skewness')
                    if (the_min < 0):
                        print('negative-values')
                        all_positive = new_attribute + 2*abs(the_min)
                    else:
                        all_positive = new_attribute

                    if np.std(all_positive)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:log')
                        log = np.log(all_positive)
                        data_2d = log.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-log')

                elif (skewness < -1):
                    print('left-skewness')
                    if (the_min < 0):
                        print('negative-values')
                        all_positive = new_attribute + abs(the_min)
                    else:
                        all_positive = new_attribute

                    if np.std(all_positive)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:sqrt')
                        sqrt = np.sqrt(all_positive)
                        data_2d = sqrt.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-sqrt')

                else:
                    if np.std(new_attribute)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:zscore')
                        zscore = stats.zscore(new_attribute)
                        data_2d = zscore.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-zscore')


            else:
                if (skewness > 1 or skewness < -1):
                    if skewness > 1:
                        print('right-skewness')
                    elif skewness < -1:
                        print('left-skewness')


                    if (the_min < 0):
                        print('negative-values')
                        all_positive = new_attribute + 2*abs(the_min)
                    else:
                        all_positive = new_attribute

                    if np.std(all_positive)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:boxcox')
                        boxcox, values = stats.boxcox(all_positive)
                        data_2d = boxcox.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-boxcox')

                else:
                    if np.std(new_attribute)== 0.0:
                        minmax = np.ones(new_partition.shape[0]*num_timestamps)
                    else:
                        print('methode:zscore')
                        zscore = stats.zscore(new_attribute)
                        data_2d = zscore.reshape(-1, 1)
                        minmax = scaler.fit_transform(data_2d).flatten()
                    if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                        print('nan-zscore')

                new_attribute = minmax

            for m in range(0,new_partition.shape[0]):
                new_partition[m,j*num_timestamps:(j+1)*num_timestamps] = minmax[m*num_timestamps:(m+1)*num_timestamps] 



       

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_4_FinalData_Concatenation_LSBZM_KnnImputation/"
name = 'Concatenation_LSBZM_KnnImputation'

lsbzm_info(4,4,concat_imputed_data,labels, data_dir, name)
