==> Part7_Classification_ML.ipynb <==
# Classification (SVM, MLPClassifier, KNN, Randon Forest) on Vector Data

### Evaluation Metrics

# float("{:.2f}".format(13.949999999999999))

def TSS(TP,TN,FP,FN):
    TSS_value = (TP / (TP + FN)) - (FP / (FP + TN))
    return TSS_value

def HSS1(TP,TN,FP,FN):
    HSS1_value = (2 * (TP * TN - FP * FN)) / ((TP + FN) * (FN + TN) + (TP + FP) * (FP + TN))
    return HSS1_value
    
def HSS2(TP,TN,FP,FN):
    HSS2_value = (2 * (TP * TN - FP * FN)) / ((TP + FP) * (FN + TN) + (TP + FN) * (FP + TN))
    return HSS2_value

def GSS(TP,TN,FP,FN):
    GSS_value = (TP - (TP + FP) * (TP + FN) / (TP + FP + FN + TN))
    return GSS_value

def Recall(TP,TN,FP,FN):
    Recall_value = (TP) / (TP + FN)
    return Recall_value

def FPR(TP,TN,FP,FN):
    fpr_value = (FP) / (FP + TN)
    return fpr_value

def Accuracy(TP,TN,FP,FN):
    accuracy_value = (TP + TN) / (TP + TN + FP + FN)
    return accuracy_value

def Precision(TP,TN,FP,FN):
    precision_value = (TP) / (TP + FP)
    return precision_value

### Reading the Data built by the Previous Notebooks (LSBZM Normalization and FPCKNN Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_2_FinalData_NewFeatures_LSBZM_KnnImputation/"
X_train_NewF_LSBZM = []
Y_train_NewF_LSBZM = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        X_train_NewF_LSBZM.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_LSBZM[i]).any() or np.isinf(X_train_NewF_LSBZM[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_LSBZM.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (MinMax Normalization and Mean Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_M = []
Y_train_NewF_M = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_MinMaxNorm_MeanImputation" + ".pkl", 'rb') as f:
        X_train_NewF_M.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_M[i]).any() or np.isinf(X_train_NewF_M[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_MinMaxNorm_MeanImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_M.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (Z-score Normalization and Mean Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_Z = []
Y_train_NewF_Z = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_ZNorm_MeanImputation" + ".pkl", 'rb') as f:
        X_train_NewF_Z.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_Z[i]).any() or np.isinf(X_train_NewF_Z[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_ZNorm_MeanImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_Z.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (MinMax Normalization and Nextvalue Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_M_N = []
Y_train_NewF_M_N = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_MinMaxNorm_NextvalueImputation" + ".pkl", 'rb') as f:
        X_train_NewF_M_N.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_M_N[i]).any() or np.isinf(X_train_NewF_M_N[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_MinMaxNorm_NextvalueImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_M_N.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (Z-score Normalization and Nextvalue Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_Z_N = []
Y_train_NewF_Z_N = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_ZNorm_NextvalueImputation" + ".pkl", 'rb') as f:
        X_train_NewF_Z_N.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_Z_N[i]).any() or np.isinf(X_train_NewF_Z_N[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_ZNorm_NextvalueImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_Z_N.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (Z-score Normalization and FPCKNN Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_1_FinalData_NewFeatures_ZM_KnnImputation/"
X_train_NewF_Z_KNN = []
Y_train_NewF_Z_KNN = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_ZNorm_KnnImputation" + ".pkl", 'rb') as f:
        X_train_NewF_Z_KNN.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_Z_KNN[i]).any() or np.isinf(X_train_NewF_Z_KNN[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_ZNorm_KnnImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_Z_KNN.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (MinMax Normalization and FPCKNN Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_1_FinalData_NewFeatures_ZM_KnnImputation/"
X_train_NewF_M_KNN = []
Y_train_NewF_M_KNN = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_MinMaxNorm_KnnImputation" + ".pkl", 'rb') as f:
        X_train_NewF_M_KNN.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_M_KNN[i]).any() or np.isinf(X_train_NewF_M_KNN[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_MinMaxNorm_KnnImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_M_KNN.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (LSBZM Normalization and Mean Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_LSBZM_M = []
Y_train_NewF_LSBZM_M = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_LSBZM_MeanImputation" + ".pkl", 'rb') as f:
        X_train_NewF_LSBZM_M.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_LSBZM_M[i]).any() or np.isinf(X_train_NewF_LSBZM_M[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_LSBZM_MeanImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_LSBZM_M.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (LSBZM Normalization and Nextvalue Imputation)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_0_FinalData_ZM_BaseLineImputation/"
X_train_NewF_LSBZM_N = []
Y_train_NewF_LSBZM_N = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_LSBZM_NextvalueImputation" + ".pkl", 'rb') as f:
        X_train_NewF_LSBZM_N.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_LSBZM_N[i]).any() or np.isinf(X_train_NewF_LSBZM_N[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_LSBZM_NextvalueImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_LSBZM_N.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (FPCKNN Imputation and Without B and C Classes)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_5_FinalData_NewFeatures_WithoutB&C_LSBZM_KnnImputation/"
X_train_NewF_LSBZM_wBandC = []
Y_train_NewF_LSBZM_wBandC = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_WithoutB&C_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        X_train_NewF_LSBZM_wBandC.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_LSBZM_wBandC[i]).any() or np.isinf(X_train_NewF_LSBZM_wBandC[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_WithoutB&C_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_LSBZM_wBandC.append(pickle.load(f))

### Reading the Data built by the Previous Notebooks (FPCKNN Imputation and Without C Class)

import pickle
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/5_6_FinalData_NewFeatures_WithoutC_LSBZM_KnnImputation/"
X_train_NewF_LSBZM_wC = []
Y_train_NewF_LSBZM_wC = []


num_partitions = 5

for i in range(0,num_partitions):
    with open(data_dir + "Partition" + str(i+1) + "_NewFeatures_WithoutC_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        X_train_NewF_LSBZM_wC.append(pickle.load(f))
    print("P"+str(i+1)+" Nan-Value: "+ str(np.isnan(X_train_NewF_LSBZM_wC[i]).any() or np.isinf(X_train_NewF_LSBZM_wC[i]).any()))
    
    with open(data_dir + "Partition" + str(i+1) + "_Labels_NewFeatures_WithoutC_LSBZM_KnnImputation" + ".pkl", 'rb') as f:
        Y_train_NewF_LSBZM_wC.append(pickle.load(f))

### Function to run the classification on different Train-Test Splits

def kfold_training(name, X_train, Y_train, training_func, num):
    kfold = np.array([[1,2],[2,3],[3,4],[4,5]])

    metrics = []
    metrics_values = np.array([])
    
    for i in range(0, num):
        train_index = kfold[i,0]
        test_index = kfold[i,1]
        metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_train[test_index-1], Y_train[test_index-1])
        count = 0
        while (metrics_values[0] <= 0.1):
            count += 1
            metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_train[test_index-1], Y_train[test_index-1])
            if count == 5:
                break
        
        metrics.append(np.append(np.append(train_index, test_index), metrics_values))
    return metrics

def kfold_training_sampling(name, X_train, Y_train, X_test, Y_test, training_func, num):
    kfold = np.array([[1,2],[2,3],[3,4],[4,5]])

    metrics = []
    metrics_values = np.array([])
    
    for i in range(0, num):
        train_index = kfold[i,0]
        test_index = kfold[i,1]
        metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_test[test_index-1], Y_test[test_index-1])
        count = 0
        while (metrics_values[0] <= 0.1):
            count += 1
            metrics_values = training_func(X_train[train_index-1], Y_train[train_index-1], X_test[test_index-1], Y_test[test_index-1])
            if count == 5:
                break
        
        metrics.append(np.append(np.append(train_index, test_index), metrics_values))
    return metrics

## Code of the Classification techniques

### SVM

# Import necessary libraries
import numpy as np
from sklearn.svm import SVC
from sklearn.metrics import confusion_matrix

def svm_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"


    # Create an SVM classifier (you can choose different kernels like 'linear', 'rbf', etc.)
    svm_classifier = SVC(kernel='rbf', C=1.0)
    svm_classifier.fit(X_train, Y_train)
    y_pred = svm_classifier.predict(X_test)
    
    
    print(str(X_train.shape)+': SVM Classifier is Done! \n')
    

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "svm_model.pkl")

    #loaded_svm_model = joblib.load(data_dir + "svm_model.pkl")
    
    return output_values

### MLP

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import numpy as np
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers
from sklearn.metrics import confusion_matrix

def mlp_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"

    # Define the MLP model
    # Define the MLP model with four hidden layers
    model = keras.Sequential([
        layers.Input(shape=(216,)),  # Input layer with 216 features
        layers.Dense(64, activation='relu'),  # Hidden layer with 64 units and ReLU activation
        layers.Dense(32, activation='relu'),  # Hidden layer with 32 units and ReLU activation
        layers.Dense(16, activation='relu'),  # Hidden layer with 16 units and ReLU activation
        layers.Dense(8, activation='relu'),  # Hidden layer with 16 units and ReLU activation
        layers.Dense(1, activation='sigmoid')  # Output layer with 1 unit and sigmoid activation (binary classification)
    ])

    # Compile the model
    model.compile(optimizer='adam', loss='binary_crossentropy', metrics=keras.metrics.PrecisionAtRecall(0.95))

    # Train the model
    model.fit(X_train, Y_train, epochs=15, batch_size=32, verbose=0)  # Adjust epochs and batch_size as needed

    y_pred = model.predict(X_test)
    threshold = 0.5  # Adjust the threshold as needed
    y_pred_binary = (y_pred > threshold).astype(int)
    
    print(str(X_train.shape)+': MLP Classifier is Done! \n')


    confusion = confusion_matrix(Y_test, y_pred_binary)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])


    #joblib.dump(classifier, data_dir + "mlp_model.pkl")

    #loaded_mlp_model = joblib.load(data_dir + "mlp_model.pkl")
    
    return output_values

## k-NN

import numpy as np
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import confusion_matrix

def knn_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"

    # Create a Gaussian Naive Bayes classifier
    nb_classifier = KNeighborsClassifier(n_neighbors=5)
    nb_classifier.fit(X_train, Y_train)
    y_pred = nb_classifier.predict(X_test)
    
    print(str(X_train.shape) + ': KNeighborsClassifier Classifier is Done! \n')

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp,tn,fp,fn)
    hss1 = HSS1(tp,tn,fp,fn)
    hss2 = HSS2(tp,tn,fp,fn)
    gss = GSS(tp,tn,fp,fn)
    recall = Recall(tp,tn,fp,fn)
    precision = Precision(tp,tn,fp,fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])

    #joblib.dump(nb_classifier, data_dir + "naive_bayes_model.pkl")

    #loaded_nb_model = joblib.load(data_dir + "naive_bayes_model.pkl")

    return output_values


### Random Forest

import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import confusion_matrix

def random_forest_model(X_train, Y_train, X_test, Y_test):
    
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/models/"

    # Create a Random Forest classifier
    # You can adjust 'n_estimators' and other parameters as needed
    rf_classifier = RandomForestClassifier(n_estimators=64, random_state=42, criterion='entropy')
    rf_classifier.fit(X_train, Y_train)
    y_pred = rf_classifier.predict(X_test)
    
    print(str(X_train.shape) + ': Random Forest Classifier is Done! \n')

    confusion = confusion_matrix(Y_test, y_pred)
    tn, fp, fn, tp = confusion.ravel()

    tss = TSS(tp, tn, fp, fn)
    hss1 = HSS1(tp, tn, fp, fn)
    hss2 = HSS2(tp, tn, fp, fn)
    gss = GSS(tp, tn, fp, fn)
    recall = Recall(tp, tn, fp, fn)
    precision = Precision(tp, tn, fp, fn)
    
    output_values = np.array([tp, fn, fp, tn, tss, hss1, hss2, gss, recall, precision])

    #joblib.dump(rf_classifier, data_dir + "random_forest_model.pkl")

    #loaded_rf_model = joblib.load(data_dir + "random_forest_model.pkl")

    return output_values


### Saving the Result of Classification on File

def save_results(reslut, name):
    data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/results/"

    with open(data_dir + name + ".pkl", 'wb') as f:
        pickle.dump(reslut, f)

## Running Experiments 

### SVM

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_LSBZM, Y_train_NewF_LSBZM, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_M, Y_train_NewF_M, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_MM_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_Z, Y_train_NewF_Z, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_ZM_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_M_N, Y_train_NewF_M_N, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_MN_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_Z_N, Y_train_NewF_Z_N, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_ZN_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_Z_KNN, Y_train_NewF_Z_KNN, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_ZKNN_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_M_KNN, Y_train_NewF_M_KNN, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_MKNN_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_LSBZM_M, Y_train_NewF_LSBZM_M, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_LSBZM_M_Results")

# SVM NewFeatures
svm_newf = kfold_training('SVM', X_train_NewF_LSBZM_N, Y_train_NewF_LSBZM_N, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_LSBZM_N_Results")

# SVM NewFeatures
svm_newf = kfold_training_sampling('SVM', X_train_NewF_LSBZM_wBandC, Y_train_NewF_LSBZM_wBandC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_WithoutB&C_LSBZM_KNN_Results")

# SVM NewFeatures
svm_newf = kfold_training_sampling('SVM', X_train_NewF_LSBZM_wC, Y_train_NewF_LSBZM_wC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, svm_model, 4)

save_results(svm_newf, "SVM_NewFeatures_WithoutC_LSBZM_KNN_Results")

### MLP

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_LSBZM, Y_train_NewF_LSBZM, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_M, Y_train_NewF_M, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_MM_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_Z, Y_train_NewF_Z, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_ZM_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_M_N, Y_train_NewF_M_N, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_MN_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_Z_N, Y_train_NewF_Z_N, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_ZN_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_Z_KNN, Y_train_NewF_Z_KNN, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_ZKNN_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_M_KNN, Y_train_NewF_M_KNN, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_MKNN_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_LSBZM_M, Y_train_NewF_LSBZM_M, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_LSBZM_M_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training('MLP', X_train_NewF_LSBZM_N, Y_train_NewF_LSBZM_N, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_LSBZM_N_Results")

# MLPClassifier NewFeatures
mlp_newf = kfold_training_sampling('MLP', X_train_NewF_LSBZM_wBandC, Y_train_NewF_LSBZM_wBandC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_WithoutB&C_LSBZM_KNN_Results")

mlp_newf = kfold_training_sampling('MLP', X_train_NewF_LSBZM_wC, Y_train_NewF_LSBZM_wC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, mlp_model, 4)

save_results(mlp_newf, "MLPClassifier_NewFeatures_WithoutC_LSBZM_KNN_Results")

### k-NN

naive_newf = kfold_training('KNN', X_train_NewF_LSBZM, Y_train_NewF_LSBZM, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_Results")

naive_newf = kfold_training('KNN', X_train_NewF_M, Y_train_NewF_M, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_MM_Results")

naive_newf = kfold_training('KNN', X_train_NewF_Z, Y_train_NewF_Z, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_ZM_Results")

naive_newf = kfold_training('KNN', X_train_NewF_M_N, Y_train_NewF_M_N, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_MN_Results")

naive_newf = kfold_training('KNN', X_train_NewF_Z_N, Y_train_NewF_Z_N, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_ZN_Results")

naive_newf = kfold_training('KNN', X_train_NewF_Z_KNN, Y_train_NewF_Z_KNN, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_ZKNN_Results")

naive_newf = kfold_training('KNN', X_train_NewF_M_KNN, Y_train_NewF_M_KNN, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_MKNN_Results")

naive_newf = kfold_training('KNN', X_train_NewF_LSBZM_M, Y_train_NewF_LSBZM_M, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_LSBZM_M_Results")

naive_newf = kfold_training('KNN', X_train_NewF_LSBZM_N, Y_train_NewF_LSBZM_N, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_LSBZM_N_Results")

naive_newf = kfold_training_sampling('KNN', X_train_NewF_LSBZM_wBandC, Y_train_NewF_LSBZM_wBandC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_WithoutB&C_LSBZM_KNN_Results")

naive_newf = kfold_training_sampling('KNN', X_train_NewF_LSBZM_wC, Y_train_NewF_LSBZM_wC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, knn_model, 4)

save_results(naive_newf, "KNN_NewFeatures_WithoutC_LSBZM_KNN_Results")

### Random Forest

forest_newf = kfold_training('RandomForest', X_train_NewF_LSBZM, Y_train_NewF_LSBZM, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_M, Y_train_NewF_M, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_MM_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_Z, Y_train_NewF_Z, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_ZM_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_M_N, Y_train_NewF_M_N, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_MN_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_Z_N, Y_train_NewF_Z_N, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_ZN_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_Z_KNN, Y_train_NewF_Z_KNN, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_ZKNN_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_M_KNN, Y_train_NewF_M_KNN, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_MKNN_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_LSBZM_M, Y_train_NewF_LSBZM_M, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_LSBZM_M_Results")

forest_newf = kfold_training('RandomForest', X_train_NewF_LSBZM_N, Y_train_NewF_LSBZM_N, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_LSBZM_N_Results")

forest_newf = kfold_training_sampling('RandomForest', X_train_NewF_LSBZM_wBandC, Y_train_NewF_LSBZM_wBandC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_WithoutB&C_LSBZM_KNN_Results")

forest_newf = kfold_training_sampling('RandomForest', X_train_NewF_LSBZM_wC, Y_train_NewF_LSBZM_wC,
                                   X_train_NewF_LSBZM, Y_train_NewF_LSBZM, random_forest_model, 4)

save_results(forest_newf, "RandomForest_NewFeatures_WithoutC_LSBZM_KNN_Results")
