==> Part13_NormalizationAfterSampling.ipynb <==
# Normalization on Concatenated Multivariate Time Series Data (2D)

## LSBZM Normalization Technique

import warnings
warnings.filterwarnings("ignore")

from sklearn.preprocessing import MinMaxScaler
scaler = MinMaxScaler()
from scipy import stats
from tqdm import tqdm

def lsbzm(start_partition, end_partition, data, labels, data_dir, first_name, last_name):

    num_timestamps= 60
    num_attributes = 24

    for i in range(start_partition-1,end_partition):
        new_partition = np.zeros((np.array(data[i]).shape[0], num_timestamps*num_attributes))
        new_partition = np.array(data[i])

        with tqdm(num_attributes) as pbar:
            for j in range(0,num_attributes):
                new_attribute = np.zeros(new_partition.shape[0]*num_timestamps)  
                
                for m in range(0,new_partition.shape[0]):
                    new_attribute[m*num_timestamps:(m+1)*num_timestamps] = new_partition[m,j*num_timestamps:(j+1)*num_timestamps]
                
                minmax = np.zeros(new_partition.shape[0]*num_timestamps)
                all_positive = np.zeros(new_partition.shape[0]*num_timestamps)

                the_min = np.min(new_attribute)
                the_max = np.max(new_attribute)
                skewness = stats.skew(new_attribute)
                
                if (the_max - the_min > 100000):
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log')

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt')

                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')
                
                
                        
                elif (the_max - the_min < 0.00001):
                    
                    if (skewness > 1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute
                            
                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            log = np.log(all_positive)
                            data_2d = log.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-log')

                    elif (skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            sqrt = np.sqrt(all_positive)
                            data_2d = sqrt.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-sqrt')

                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')

                        
                else:
                    if (skewness > 1 or skewness < -1):
                        if (the_min < 0):
                            all_positive = new_attribute + 2*abs(the_min)
                        else:
                            all_positive = new_attribute

                        if np.std(all_positive)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            boxcox, values = stats.boxcox(all_positive)
                            data_2d = boxcox.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-boxcox')
                            
                    else:
                        if np.std(new_attribute)== 0.0:
                            minmax = np.ones(new_partition.shape[0]*num_timestamps)
                        else:
                            zscore = stats.zscore(new_attribute)
                            data_2d = zscore.reshape(-1, 1)
                            minmax = scaler.fit_transform(data_2d).flatten()
                        if (np.isnan(minmax).any() or np.isinf(minmax).any()):
                            print('nan-zscore')

                    new_attribute = minmax

                for m in range(0,new_partition.shape[0]):
                    new_partition[m,j*num_timestamps:(j+1)*num_timestamps] = minmax[m*num_timestamps:(m+1)*num_timestamps] 
                 

                pbar.update(1)

       

            with open(data_dir + first_name + "Partition" + str(i+1) + last_name + ".pkl", 'wb') as f:
                pickle.dump(new_partition, f)
            with open(data_dir + first_name + "Partition" + str(i+1) + "_Labels" + last_name +".pkl", 'wb') as f:
                pickle.dump(labels[i], f)

### Reading the Data built by the Previous Notebooks (RUS_Tomek_Smote)

import pickle
import pandas as pd
import numpy as np

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_Smote_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_Smote_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'RUS_Tomek_Smote_'
last_name = '_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (RUS_Tomek_Adasyn)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_Adasyn_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_Adasyn_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'RUS_Tomek_Adasyn_'
last_name = '_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (Smote)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "Smote_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "Smote_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'Smote_'
last_name = '_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (Adasyn)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "Adasyn_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "Adasyn_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'Adasyn_'
last_name = '_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (GNI)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "GaussianNoise_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "GaussianNoise_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'GaussianNoise_'
last_name = '_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (GNI, Tomek, RUS)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_GNI_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_GNI_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'RUS_Tomek_GNI_'
last_name = '_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (TimeGAN)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_2_OSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "TimeGAN_" +"Partition" + str(i+1) + "_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "timeGAN_" + "Partition" + str(i+1) + "_Labels_OSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_4_FinalData_OSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'TimeGAN_'
last_name = '_OSampling_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)

### Reading the Data built by the Previous Notebooks (TimeGAN, Tomek, RUS)

import pickle
import pandas as pd

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/6_1_OUSampling_Concatenation_KnnImputation/"

X_train_array = []
Y_train_array = []

num_partitions = 5
for i in range(0, num_partitions):
    with open(data_dir + "RUS_Tomek_TimeGAN_" +"Partition" + str(i+1) + "_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        X_train_array.append(pickle.load(f))
    with open(data_dir + "RUS_Tomek_TimeGAN_" + "Partition" + str(i+1) + "_Labels_OUSampling_WithoutC_Concatenation_KnnImputation" +".pkl", 'rb') as f:
        Y_train_array.append(pickle.load(f))

data_dir = "/Users/samskanderi/Documents/Research_Project/SWANSF/code/8_3_FinalData_OUSampling_Concatenation_LSBZM_KnnImputation/"
first_name = 'RUS_Tomek_TimeGAN_'
last_name = '_WithoutC_Concatenation_LSBZM_KnnImputation'

lsbzm(1,5,X_train_array,Y_train_array, data_dir, first_name, last_name)
